@@ -1,4 +1,4 @@
-package org.hyperskill.flashcards;
+package org.hyperskill.flashcards.configuration;
 
 public enum ActionsEnum {
     ADD, REMOVE, IMPORT, EXPORT, ASK, EXIT, LOG, HARDEST_CARD, RESET_STATS, DEFAULT
